requirements.txt
tensorflow
requests
streamlit
pillow
requests
selenium